merge upstream release 6.1.0

+++ src/main/io/osd_canvas.c
@@ -279,7 +279,7 @@ static void osdDrawArtificialHorizonShapes(displayCanvas_t *canvas, float pitchA
         itoa(absLevel, buf, 10);
         int pos = level * pixelsPerDegreeLevel;
         int charY = 9 - pos * 2;
-        int cx = (absLevel >= 100 ? -1.5f : -1.0) * canvas->gridElementWidth;
+        int cx = (absLevel >= 100 ? -1.5f : -1.0f) * canvas->gridElementWidth;
         int px = cx + (pitchOffset + pos) * sx * 2;
         int py = -charY - (pitchOffset + pos) * (1 - sy) * 2;
         displayCanvasDrawString(canvas, px, py, buf, 0);
@@ -438,7 +438,7 @@ void osdCanvasDrawArtificialHorizon(displayPort_t *display, displayCanvas_t *can
 
 void osdCanvasDrawHeadingGraph(displayPort_t *display, displayCanvas_t *canvas, const osdDrawPoint_t *p, int heading)
 {
-    static const uint16_t graph[] = {
+    static const uint8_t graph[] = {
         SYM_HEADING_W,
         SYM_HEADING_LINE,
         SYM_HEADING_DIVIDED_LINE,
@@ -517,7 +517,7 @@ static int32_t osdCanvasSidebarGetValue(osd_sidebar_scroll_e scroll)
         case OSD_SIDEBAR_SCROLL_SPEED:
             {
 #if defined(USE_GPS)
-                int speed = osdGetSpeedFromSelectedSource();
+                int16_t speed = osdGetSpeedFromSelectedSource();
                 switch ((osd_unit_e)osdConfig()->units) {
                     case OSD_UNIT_UK:
                         FALLTHROUGH;
