+++ src/main/io/osd_common.h
@@ -45,7 +45,7 @@ typedef struct {
 
 PG_DECLARE(osdCommonConfig_t, osdCommonConfig);
 
-int osdGetSpeedFromSelectedSource(void);
+int16_t osdGetSpeedFromSelectedSource(void);
 
 #endif // defined(USE_OSD) || defined(USE_DJI_HD_OSD)
